# mcp-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GigRadar — Fiverr Market Intelligence API",
    "description": "Scrape Fiverr gigs, analyze market data, compare sellers, filter results, and export reports. Use this API to research freelancer pricing, seller demographics, and competitive positioning on Fiverr.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3456",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/datasets/{id}": {
      "get": {
        "operationId": "getDataset",
        "summary": "Get a dataset by ID",
        "description": "Retrieve the full gig data for a specific dataset, with pagination support.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 }, "description": "Max gigs to return" },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 }, "description": "Offset for pagination" }
        ],
        "responses": {
          "200": {
            "description": "Dataset with gig data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "title": { "type": "string" },
                    "totalGigs": { "type": "integer" },
                    "showing": { "type": "string" },
                    "gigs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "title": { "type": "string" },
                          "seller": { "type": "string" },
                          "sellerCountry": { "type": "string" },
                          "sellerLevel": { "type": "string" },
                          "price": { "type": "number" },
                          "deliveryDays": { "type": "integer" },
                          "rating": { "type": "number" },
                          "reviewCount": { "type": "integer" },
                          "isPromoted": { "type": "boolean" },
                          "url": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteDataset",
        "summary": "Delete a dataset",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Dataset deleted" }
        }
      }
    },
...
    "/api/datasets/{id}/export": {
      "get": {
        "operationId": "exportDataset",
        "summary": "Export a dataset",
        "description": "Export a dataset in CSV, JSON, or Markdown report format.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["csv", "json", "markdown"], "default": "json" } }
        ],
        "responses": {
          "200": { "description": "Exported data in the requested format" }
        }
      }
    },
...
  }
}